fastapi_server/pyproject.toml
[project]
name = "fastapi-server"
version = "0.1.0"
description = "FastAPI Hello World Server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "vault-file",
    "app-yaml-config",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fastapi-server"
version = "0.1.0"
description = "FastAPI Hello World Server"
authors = ["Developer <dev@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.104.0"
uvicorn = {version = ">=0.24.0", extras = ["standard"]}
python-dotenv = "^1.0.0"
deepmerge = "^1.1.0"
vault-file = {path = "../packages_py/vault_file", develop = true}
app-yaml-config = {path = "../packages_py/app_yaml_config", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
httpx = ">=0.25.0"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app"]
